Move buildHandles function to correct util file
